Fix JavaFX bug
Reorganised files into proper packages for ease of developing
