+++ data/jarvis.txt
@@ -1,5 +1,5 @@
-[T][ ] cook dinner
 [D][ ] finish essay  (Due: Feb 22 2023)
 [D][ ] read lecture  (Due: Feb 23 2023)
 [E][ ] travel  (Start: Feb 25 2023 | End: Feb 28 2023)
 [T][ ] feed the fish
+[T][ ] cook dinner
