+++ src/main/java/exception/NoTaskFoundException.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package exception;
 
 /**
  * NoTaskFoundException class.
