+++ src/main/java/exception/NotValidDateException.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package exception;
 
 /**
  * NotValidDateException class.
