+++ src/main/java/gui/Launcher.java
@@ -1,0 +1,13 @@
+package gui;
+
+import jarvis.Duke;
+import javafx.application.Application;
+
+/**
+ * Launcher to launch Duke.
+ */
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Duke.class, args);
+    }
+}
