+++ src/main/java/jarvis/DialogBox.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package jarvis;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
