+++ src/main/java/jarvis/Storage.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package jarvis;
 
 import java.io.File;
 import java.io.FileNotFoundException;
