+++ src/main/java/jarvis/ToDoList.java
@@ -1,4 +1,9 @@
-package seedu.duke;
+package jarvis;
+
+import task.Deadline;
+import task.Event;
+import task.Task;
+import task.Todo;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
