+++ src/main/java/jarvis/Ui.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package jarvis;
 
 import java.util.Scanner;
