--- src/main/java/seedu/duke/Duke.java
@@ -1,184 +1,0 @@
-package seedu.duke;
-
-import java.io.FileNotFoundException;
-import java.time.LocalDate;
-
-import javafx.application.Application;
-import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.control.ScrollPane;
-import javafx.scene.control.TextField;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.Region;
-import javafx.scene.layout.VBox;
-import javafx.stage.Stage;
-
-
-/**
- * The Duke chatbot.
- */
-public class Duke extends Application {
-
-    private static String dataPath = "./data/jarvis.txt";
-    private final Storage storage;
-    private ToDoList todolist;
-    private Ui ui;
-    private final Parser parser;
-    private boolean isBye;
-    private ScrollPane scrollPane;
-    private VBox dialogContainer;
-    private TextField userInput;
-    private Button sendButton;
-    private Scene scene;
-    private Image user = new Image(this.getClass().getResourceAsStream("/images/TonyStark.png"));
-    private Image jarvis = new Image(this.getClass().getResourceAsStream("/images/jarvis.png"));
-    /**
-     * Duke Constructor.
-     */
-    public Duke() {
-        storage = new Storage(dataPath);
-        todolist = new ToDoList();
-        ui = new Ui();
-        parser = new Parser();
-        isBye = false;
-
-        // Print previous data
-        try {
-            storage.loadTasks(todolist);
-        } catch (FileNotFoundException e) {
-            System.out.println("File not found");
-        }
-    }
-
-    /**
-     * Starts the Duke chatbot.
-     */
-    public void run() {
-        AsciiArt asciiArt = new AsciiArt();
-        asciiArt.printArt();
-
-        System.out.println("\nPlease enter a command Mr Stark.");
-    }
-
-    @Override
-    public void start(Stage stage) {
-        //Step 1. Setting up required components
-
-        //The container for the content of the chat to scroll.
-        scrollPane = new ScrollPane();
-        dialogContainer = new VBox();
-        scrollPane.setContent(dialogContainer);
-
-        userInput = new TextField();
-        sendButton = new Button("Send");
-
-        AnchorPane mainLayout = new AnchorPane();
-        mainLayout.getChildren().addAll(scrollPane, userInput, sendButton);
-
-        scene = new Scene(mainLayout);
-
-        stage.setScene(scene);
-        stage.show();
-
-        //Step 2. Formatting the window to look as expected
-        stage.setTitle("J.A.R.V.I.S");
-        stage.setResizable(false);
-        stage.setMinHeight(800.0);
-        stage.setMinWidth(600.0);
-
-        mainLayout.setPrefSize(600.0, 800.0);
-
-        scrollPane.setPrefSize(585, 735);
-        scrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
-        scrollPane.setVbarPolicy(ScrollPane.ScrollBarPolicy.ALWAYS);
-
-        scrollPane.setVvalue(1.0);
-        scrollPane.setFitToWidth(true);
-
-        // You will need to import `javafx.scene.layout.Region` for this.
-        dialogContainer.setPrefHeight(Region.USE_COMPUTED_SIZE);
-
-        userInput.setPrefWidth(525.0);
-
-        sendButton.setPrefWidth(55.0);
-
-        AnchorPane.setTopAnchor(scrollPane, 1.0);
-
-        AnchorPane.setBottomAnchor(sendButton, 1.0);
-        AnchorPane.setRightAnchor(sendButton, 1.0);
-
-        AnchorPane.setLeftAnchor(userInput , 1.0);
-        AnchorPane.setBottomAnchor(userInput, 1.0);
-
-        //Part 3. Add functionality to handle user input.
-        sendButton.setOnMouseClicked((event) -> {
-            handleUserInput();
-        });
-
-        userInput.setOnAction((event) -> {
-            handleUserInput();
-        });
-
-        //Scroll down to the end every time dialogContainer's height changes.
-        dialogContainer.heightProperty().addListener((observable) -> scrollPane.setVvalue(1.0));
-    }
-
-    /**
-     * Iteration 1:
-     * Creates a label with the specified text and adds it to the dialog container.
-     *
-     * @param text String containing text to add
-     * @return a label with the specified text that has word wrap enabled.
-     */
-    private Label getDialogLabel(String text) {
-        // You will need to import `javafx.scene.control.Label`.
-        Label textToAdd = new Label(text);
-        textToAdd.setWrapText(true);
-
-        return textToAdd;
-    }
-
-    /**
-     * Iteration 2:
-     * Creates two dialog boxes, one echoing user input and the other containing Duke's reply and then appends them to
-     * the dialog container. Clears the user input after processing.
-     */
-    private void handleUserInput() {
-        Label userText = new Label(userInput.getText());
-        Label dukeText = new Label(getResponse(userInput.getText()));
-        dialogContainer.getChildren().addAll(
-                DialogBox.getUserDialog(userText, new ImageView(user)),
-                DialogBox.getDukeDialog(dukeText, new ImageView(jarvis))
-        );
-        userInput.clear();
-    }
-
-    /**
-     * You should have your own function to generate a response to user input.
-     * Replace this stub with your completed method.
-     */
-    private String getResponse(String input) {
-        assert input != null : "No input given";
-        return parser.parse(input, todolist, storage);
-    }
-
-    /**
-     * Returns the current date.
-     *
-     * @return the current date
-     */
-    public static LocalDate getCurrDate() {
-        return LocalDate.now();
-    }
-
-    /**
-     * main.
-     * @param args main args
-     */
-    public static void main(String[] args) {
-        Application.launch(Duke.class, args);
-    }
-}
