+++ src/main/java/task/Deadline.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
