+++ src/test/META-INF/MANIFEST.MF
