+++ src/test/java/jarvis/TodoTest.java
@@ -1,8 +1,9 @@
-package seedu.duke;
+package jarvis;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
+import task.Todo;
 
 public class TodoTest {
